refactor: step1 역할 분리 피드백 수정 -Delimiter   splitWithDelimiter 로직 내부에서 수정 -Operands   int Arr -> List<Operand>를 담는 생성자 추가
