+++ src/main/java/calculator/domain/Operands.java
@@ -1,11 +1,13 @@
 /*
-* Operand 집합을 포장하는 클래스
-* */
-package Calculator.domain;
+ * Operand 집합을 포장하는 클래스
+ * */
+package calculator.domain;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.function.BinaryOperator;
+import java.util.stream.Collectors;
 
 public class Operands {
     private List<Operand> operands;
@@ -13,6 +15,14 @@ public class Operands {
     public Operands(List<Operand> operands) {
         this.operands = operands;
     }
+
+    public Operands(int[] arr) {
+        this.operands = Arrays.stream(arr)
+                .mapToObj(Operand::new)
+                .filter(Operand::checkOperand)
+                .collect(Collectors.toList());
+    }
+
 
     public Operand operateAll(BinaryOperator<Integer> operator) {
         Operand result = new Operand(0);
